+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/NftDao.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface NftDao {
-    Optional<Nft> create(long nftId, String contractAddr, String nftName, String chain, MultipartFile image, long idOwner, String collection, String description, String[] properties);
+    Optional<Nft> create(int nftId, String contractAddr, String nftName, String chain, MultipartFile image, int idOwner, String collection, String description, String[] properties);
 
     Optional<Nft> getNFTById(String nftId);
 
@@ -18,9 +18,9 @@ public interface NftDao {
 
     List<Publication> getAllPublicationsByUser(int page, User user, User currentUser, boolean onlyFaved, boolean onlyOnSale);
 
-    long getAmountPublications(String status, String category, String chain, BigDecimal minPrice, BigDecimal maxPrice, String search);
+    int getAmountPublications(String status, String category, String chain, BigDecimal minPrice, BigDecimal maxPrice, String search);
 
-    void updateOwner(long nftId, long idBuyer);
+    void updateOwner(int nftId, int idBuyer);
 
     void delete(String productId);
 }
