+++ interfaces/src/main/java/ar/edu/itba/paw/service/MailingService.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 public interface MailingService {
 
     @Async
-    void sendOfferMail(String bidderEmail, String sellerMail, String nftName, long nftId, String nftAddress, BigDecimal nftPrice, byte[] image);
+    void sendOfferMail(String bidderEmail, String sellerMail, String nftName, int nftId, String nftAddress, BigDecimal nftPrice, byte[] image);
 
     @Async
     void sendRegisterMail(String userEmail, String username);
