+++ service/src/main/java/ar/edu/itba/paw/service/SellOrderServiceImpl.java
@@ -24,22 +24,22 @@ public class SellOrderServiceImpl implements SellOrderService {
     }
 
     @Override
-    public Optional<SellOrder> getOrderById(long id) {
+    public Optional<SellOrder> getOrderById(int id) {
         return sellOrderDao.getOrderById(id);
     }
 
     @Override
-    public void update(long id, String category, BigDecimal price) {
+    public void update(int id, String category, BigDecimal price) {
         sellOrderDao.update(id, category, price);
     }
 
     @Override
-    public void delete(long id) {
+    public void delete(int id) {
         sellOrderDao.delete(id);
     }
 
     @Override
-    public long getNftWithOrder(String id) {
+    public int getNftWithOrder(String id) {
         return sellOrderDao.getNftWithOrder(id);
     }
 }
